Created new Climber FSM
