+++ src/main/java/frc/robot/Robot.java
@@ -7,7 +7,7 @@ package frc.robot;
 import edu.wpi.first.wpilibj.TimedRobot;
 
 // Systems
-import frc.robot.systems.TraversalFSM;
+import frc.robot.systems.ClimberFSM;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
@@ -17,7 +17,7 @@ public class Robot extends TimedRobot {
 	private TeleopInput input;
 
 	// Systems
-	private TraversalFSM fsmSystem;
+	private ClimberFSM fsmSystem;
 
 	/**
 	 * This function is run when the robot is first started up and should be used for any
@@ -29,7 +29,7 @@ public class Robot extends TimedRobot {
 		input = new TeleopInput();
 
 		// Instantiate all systems here
-		fsmSystem = new TraversalFSM();
+		fsmSystem = new ClimberFSM();
 	}
 
 	@Override
