+++ src/main/java/frc/robot/systems/ClimberFSM.java
@@ -1,9 +1,15 @@
 package frc.robot.systems;
 
 // WPILib Imports
+import edu.wpi.first.wpilibj.DoubleSolenoid;
+import edu.wpi.first.wpilibj.PneumaticsModuleType;
+import edu.wpi.first.wpilibj.DoubleSolenoid.Value;
+import edu.wpi.first.wpilibj.Timer;
+//import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 // Third party Hardware Imports
 import com.revrobotics.CANSparkMax;
+import com.revrobotics.SparkMaxLimitSwitch;
 
 // Robot Imports
 import frc.robot.TeleopInput;
@@ -14,17 +20,30 @@ public class ClimberFSM {
 	// FSM state definitions
 	public enum FSMState {
 		IDLE,
-
+		EXTEND_TO_MAX,
+		IDLE_MAX_EXTENDED,
+		RETRACTING_TO_MIN,
+		CLIMBER_ARM_HANG,
+		ATTACH_STATIC_HANG,
+		STATIC_HANG,
+		PNEUMATIC_ACTIVATE,
+		IDLE_TILT,
+		EXTENDING_TILT
 	}
-
-	private static final float MOTOR_RUN_POWER = 0.1f;
-
-	/* ======================== Private variables ======================== */
+	private static final float ARM_MOTOR_EXTEND_POWER = 0.1f;
+	private static final float ARM_MOTOR_RETRACT_POWER = 0.1f;
+	private static final double MAX_ENCODER = 50;
+	private static int cycleCount = 0;
+	private static final int MAX_CYCLE = 4;
 	private FSMState currentState;
 
 	// Hardware devices should be owned by one and only one system. They must
 	// be private to their owner system and may not be used elsewhere.
-	private CANSparkMax exampleMotor;
+	private CANSparkMax armMotor;
+	private DoubleSolenoid armSolenoid;
+	private Timer pneumaticTimer;
+	private SparkMaxLimitSwitch armLimitSwitchExtend;
+	private SparkMaxLimitSwitch armLimitSwitchRetract;
 
 	/* ======================== Constructor ======================== */
 	/**
@@ -32,12 +51,19 @@ public class ClimberFSM {
 	 * one-time initialization or configuration of hardware required. Note
 	 * the constructor is called only once when the robot boots.
 	 */
-	public FSMSystem() {
-		// Perform hardware init
-		exampleMotor = new CANSparkMax(HardwareMap.CAN_ID_SPARK_SHOOTER,
-										CANSparkMax.MotorType.kBrushless);
-
-		// Reset state machine
+	public ClimberFSM() {
+		armMotor = new CANSparkMax(HardwareMap.CAN_ID_SPARK_CLIMBER,
+			CANSparkMax.MotorType.kBrushless);
+		armSolenoid = new DoubleSolenoid(PneumaticsModuleType.CTREPCM,
+			HardwareMap.PCM_CHANNEL_ARM_CYLINDER_EXTEND,
+			HardwareMap.PCM_CHANNEL_ARM_CYLINDER_RETRACT);
+		pneumaticTimer = new Timer();
+		armLimitSwitchExtend = armMotor.getForwardLimitSwitch(
+								SparkMaxLimitSwitch.Type.kNormallyClosed);
+		armLimitSwitchRetract = armMotor.getReverseLimitSwitch(
+								SparkMaxLimitSwitch.Type.kNormallyClosed);
+		armLimitSwitchExtend.enableLimitSwitch(true);
+		armLimitSwitchRetract.enableLimitSwitch(true);
 		reset();
 	}
 
@@ -58,7 +84,11 @@ public class ClimberFSM {
 	 * Ex. if the robot is enabled, disabled, then reenabled.
 	 */
 	public void reset() {
-		currentState = FSMState.START_STATE;
+		currentState = FSMState.IDLE;
+		armMotor.getEncoder().setPosition(0);
+		cycleCount = 0;
+		armSolenoid.set(Value.kReverse); // reset piston
+		pneumaticTimer.reset();
 
 		// Call one tick of update to ensure outputs reflect start state
 		update(null);
@@ -71,18 +101,46 @@ public class ClimberFSM {
 	 */
 	public void update(TeleopInput input) {
 		switch (currentState) {
-			case START_STATE:
-				handleStartState(input);
+			case IDLE:
+				handleIdleState(input);
 				break;
-
-			case OTHER_STATE:
-				handleOtherState(input);
+			case EXTEND_TO_MAX:
+				handleExtendToMaxState(input);
 				break;
-
+			case IDLE_MAX_EXTENDED:
+				handleIdleMaxExtendedState(input);
+				break;
+			case RETRACTING_TO_MIN:
+				handleRetractingToMinState(input);
+				break;
+			case CLIMBER_ARM_HANG:
+				handleClimberArmHangState(input);
+				break;
+			case ATTACH_STATIC_HANG:
+				handleAttachStaticHangState(input);
+				break;
+			case STATIC_HANG:
+				handleStaticHangState(input);
+				break;
+			case PNEUMATIC_ACTIVATE:
+				handlePneumaticActivateState(input);
+				break;
+			case IDLE_TILT:
+				handleIdleTiltState(input);
+				break;
+			case EXTENDING_TILT:
+				handleExtendingTiltState(input);
+				break;
 			default:
 				throw new IllegalStateException("Invalid state: " + currentState.toString());
 		}
-		currentState = nextState(input);
+
+		FSMState state = nextState(input);
+		if (currentState != state) {
+			System.out.println(state);
+		}
+		currentState = state;
+
 	}
 
 	/* ======================== Private methods ======================== */
@@ -96,37 +154,170 @@ public class ClimberFSM {
 	 * @return FSM state for the next iteration
 	 */
 	private FSMState nextState(TeleopInput input) {
+		if (input == null) {
+			return FSMState.IDLE;
+		}
 		switch (currentState) {
-			case START_STATE:
-				if (input != null) {
-					return FSMState.OTHER_STATE;
-				} else {
-					return FSMState.START_STATE;
+			case IDLE:
+				if (input.isClimberButtonPressed()) {
+					//if climber button is pressed, continue to next state
+					return FSMState.EXTEND_TO_MAX;
 				}
-
-			case OTHER_STATE:
-				return FSMState.OTHER_STATE;
+				//else, stay in idle
+				return FSMState.IDLE;
+			case EXTEND_TO_MAX:
+				if (getFirstCondition()) {
+					return FSMState.IDLE_MAX_EXTENDED;
+				}
+				return FSMState.EXTEND_TO_MAX;
+			case IDLE_MAX_EXTENDED:
+				if (!input.isClimberButtonPressed()) {
+					//climber button not pressed, stay in idle
+					return FSMState.IDLE_MAX_EXTENDED;
+				}
+				//climber button pressed, return to previous state
+				return FSMState.RETRACTING_TO_MIN;
+			case RETRACTING_TO_MIN:
+				if (getSecondCondition()) {
+					//if climber button is pressed stay in current state
+					//drivers need to release the button
+					return FSMState.CLIMBER_ARM_HANG;
+				}
+				//button is released continue to next state
+				return FSMState.RETRACTING_TO_MIN;
+			case CLIMBER_ARM_HANG:
+				if (!input.isClimberButtonPressed()) {
+					//button needs to be repressed to continue to next state, if not stay in IDLE2
+					return FSMState.CLIMBER_ARM_HANG;
+				}
+				//button is pressed so continue to next state
+				armMotor.getEncoder().setPosition(0);
+				return FSMState.ATTACH_STATIC_HANG;
+			case ATTACH_STATIC_HANG:
+				if (armMotor.getEncoder().getPosition() >= MAX_ENCODER) {
+					cycleCount++;
+					return FSMState.STATIC_HANG;
+				}
+				return FSMState.ATTACH_STATIC_HANG;
+			case STATIC_HANG:
+				if (!input.isClimberButtonPressed() || cycleCount == MAX_CYCLE) {
+					//climber button not pressed, stay in idle
+					return FSMState.STATIC_HANG;
+				}
+				//climber button pressed return to previous state
+				pneumaticTimer.reset();
+				pneumaticTimer.start();
+				return FSMState.PNEUMATIC_ACTIVATE;
+			case PNEUMATIC_ACTIVATE:
+				if (pneumaticTimer.hasElapsed(1)) {
+					return FSMState.IDLE_TILT;
+				}
+				return FSMState.PNEUMATIC_ACTIVATE;
+			case IDLE_TILT:
+				if (!input.isClimberButtonPressed()) {
+					//climber button not pressed, stay idle
+					return FSMState.IDLE_TILT;
+				}
+				//climber button pressed, go to previous state
+				return FSMState.EXTENDING_TILT;
+			case EXTENDING_TILT:
+				if (!getFirstCondition() && input.isClimberButtonPressed()) {
+					return FSMState.EXTENDING_TILT;
+				} else if (!input.isClimberButtonPressed()) {
+					return FSMState.IDLE_TILT;
+				} else {
+					return FSMState.IDLE_MAX_EXTENDED;
+				}
 
 			default:
 				throw new IllegalStateException("Invalid state: " + currentState.toString());
 		}
-	}
 
+	}
 	/* ------------------------ FSM state handlers ------------------------ */
 	/**
 	 * Handle behavior in START_STATE.
 	 * @param input Global TeleopInput if robot in teleop mode or null if
 	 *        the robot is in autonomous mode.
 	 */
-	private void handleStartState(TeleopInput input) {
-		exampleMotor.set(0);
+	private void handleIdleState(TeleopInput input) {
+		armMotor.set(0);
+		armSolenoid.set(Value.kReverse);
 	}
 	/**
-	 * Handle behavior in OTHER_STATE.
+	 * Handle behavior in first retract state.
 	 * @param input Global TeleopInput if robot in teleop mode or null if
 	 *        the robot is in autonomous mode.
 	 */
-	private void handleOtherState(TeleopInput input) {
-		exampleMotor.set(MOTOR_RUN_POWER);
+
+	private void handleExtendToMaxState(TeleopInput input) {
+		armMotor.set(ARM_MOTOR_EXTEND_POWER);
+		armSolenoid.set(Value.kReverse);
 	}
+
+	private void handleIdleMaxExtendedState(TeleopInput input) {
+		armMotor.set(0);
+		armSolenoid.set(Value.kReverse);
+	}
+
+	private void handleRetractingToMinState(TeleopInput input) {
+		armMotor.set(ARM_MOTOR_RETRACT_POWER);
+		armSolenoid.set(Value.kReverse);
+	}
+
+	private void handleClimberArmHangState(TeleopInput input) {
+		armMotor.set(0);
+		armSolenoid.set(Value.kReverse);
+	}
+
+
+	private void handleAttachStaticHangState(TeleopInput input) {
+		armMotor.set(ARM_MOTOR_EXTEND_POWER);
+		armSolenoid.set(Value.kReverse);
+	}
+
+	private void handleStaticHangState(TeleopInput input) {
+		armMotor.set(0);
+		armSolenoid.set(Value.kReverse);
+	}
+
+	private void handlePneumaticActivateState(TeleopInput input) {
+		armSolenoid.set(Value.kForward);
+		armMotor.set(0);
+	}
+
+	private void handleIdleTiltState(TeleopInput input) {
+		armMotor.set(0);
+		armSolenoid.set(Value.kForward);
+	}
+
+	private void handleExtendingTiltState(TeleopInput input) {
+		armMotor.set(ARM_MOTOR_EXTEND_POWER);
+		armSolenoid.set(Value.kForward);
+	}
+
+	private boolean getFirstCondition() {
+		return armLimitSwitchExtend.isPressed();
+	}
+
+	private boolean getSecondCondition() {
+		return armLimitSwitchRetract.isPressed();
+	}
+
+	// private void updateDashboard(TeleopInput input) {
+	// 	if (input == null) {
+	// 		SmartDashboard.putBoolean("Climber Button Pressed", false);
+	// 	} else {
+	// 		SmartDashboard.putBoolean("Climber Button Pressed", input.isClimberButtonPressed());
+	// 	}
+	// 	SmartDashboard.putNumber("Cycle Count", cycleCount);
+	// 	SmartDashboard.putNumber("Motor Encoder Position", armMotor.getEncoder().getPosition());
+	// 	SmartDashboard.putBoolean("Extension Limit Switch", armLimitSwitchFirst.isPressed());
+	// 	SmartDashboard.putBoolean("Retraction Limit Switch", armLimitSwitchSecond.isPressed());
+	// 	SmartDashboard.putNumber("Motor Power", armMotor.get());
+	// 	SmartDashboard.putNumber("Pneumatic Timer", pneumaticTimer.get());
+	// 	SmartDashboard.putString("Current State", currentState + "");
+	// 	SmartDashboard.putBoolean("Solenoid Extended", armSolenoid.get().equals(Value.kForward));
+	// }
+
 }
